requirements.txt
langchain
langchain-ollama
langgraph
langgraph-supervisor
langchain-mcp-adapters
python-dotenv
jupyterlab